🛠 fix: 개인 카드 조회 useYN 컬럼이 Default 값 부여

See merge request s10-bigdata-dist-sub2/S10P22D110!95

+++ backend/src/main/java/com/ssafy/pickachu/domain/cards/personalcards/repository/PersonalCardsRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 
 public interface PersonalCardsRepository extends JpaRepository<PersonalCards, Long>, PersonalCardsRepositoryQuerydsl {
 
-    Optional<PersonalCards> findPersonalCardsByUserIdAndCardsId(int userId, String cardsId);
+    Optional<PersonalCards> findPersonalCardsByUserIdAndCardsId(long userId, String cardsId);
 
     List<PersonalCards> findAllByUserIdAndUseYN(long userId, String userYN);
